`releases-tab` - Drop feature
