+++ .editorconfig
@@ -10,3 +10,6 @@ insert_final_newline = true
 [*.yml]
 indent_style = space
 indent_size = 2
+
+[eslint-suppressions.json]
+insert_final_newline = false
