+++ source/feature-renames.json
@@ -1,6 +1,5 @@
 {
 	"scrollable-code-and-blockquote": "scrollable-areas",
-	"add-releases-tab": "releases-tab",
 	"add-selection-in-new-tab": "selection-in-new-tab",
 	"add-profile-hotkey": "profile-hotkey",
 	"add-download-folder-button": "download-folder-button",
