+++ source/refined-github.ts
@@ -35,7 +35,6 @@ import './features/sticky-comment-header.js';
 // Disableable features
 import './features/useful-not-found-page.js';
 import './features/more-dropdown-links.js';
-import './features/releases-tab.js';
 import './features/one-key-formatting.js';
 import './features/tab-to-indent.js';
 import './features/hide-navigation-hover-highlight.js';
